[kernel] E: Refactor IPC subsystem
Refactor the IPC subsystem by splitting the monolithic kcall.rs into
dedicated send.rs and recv.rs modules. This prepares the codebase for
introducing push/pull kernel calls by establishing a cleaner module
structure.
