+++ src/kernel/src/ipc/mod.rs
@@ -5,15 +5,14 @@
 //  Modules
 //==================================================================================================
 
-mod kcall;
 mod mbx;
+mod recv;
+mod send;
 
 //==================================================================================================
 //  Exports
 //==================================================================================================
 
-pub use kcall::{
-    recv,
-    send,
-};
 pub use mbx::Mailbox;
+pub use recv::recv;
+pub use send::send;
