+++ src/kernel/src/ipc/recv.rs
@@ -1,0 +1,44 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//==================================================================================================
+// Imports
+//==================================================================================================
+
+use crate::{
+    event::EventManager,
+    pm::{
+        self,
+        ProcessManager,
+        SleepError,
+    },
+};
+use ::sys::{
+    ipc::Message,
+    pm::{
+        ProcessIdentifier,
+        ThreadIdentifier,
+    },
+};
+
+//==================================================================================================
+// Standalone Functions
+//==================================================================================================
+
+pub unsafe fn recv(
+    tid: ThreadIdentifier,
+    pid: ProcessIdentifier,
+    msg: usize,
+) -> Result<(), SleepError> {
+    if pid != ProcessIdentifier::INITD {
+        trace!("pid={:?}", pid);
+    }
+
+    match EventManager::wait(tid, pid) {
+        Ok(message) => {
+            pm::copy_to_user(ProcessManager::get_mut(), pid, msg as *mut Message, &message)
+                .map_err(SleepError::Generic)
+        },
+        Err(sleep_error) => Err(sleep_error),
+    }
+}
